# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StreamSphere Go API
  version: 1.0.0
  description: Contratos principales expuestos por el API Gateway.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VideoInput:
      type: object
      required: [channelId, title]
      properties:
        channelId: { type: string, format: uuid }
        title: { type: string, minLength: 3, maxLength: 150 }
        description: { type: string }
        category: { type: string }
        tags:
          type: array
          items: { type: string }
        visibility:
          type: string
          enum: [PUBLIC, PRIVATE, UNLISTED]
paths:
...
  /api/channels:
    post:
      summary: Crear canal y convertir al usuario en creador
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        '201': { description: Canal creado y JWT actualizado }
  /api/videos:
    post:
      summary: Crear metadatos de video
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VideoInput' }
      responses:
        '201': { description: Video creado en DRAFT }
...
  /api/playback/videos/{id}/manifest:
    get:
      summary: Obtener URL de reproducción
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: URL temporal de reproducción }
        '409': { description: Video aún no disponible }
  /api/playback/events:
    post:
      summary: Registrar progreso de reproducción
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [videoId, eventType]
              properties:
                videoId: { type: string, format: uuid }
                eventType: { type: string, enum: [STARTED, PROGRESS, COMPLETED] }
                second: { type: integer, minimum: 0 }
      responses:
        '201': { description: Evento registrado }
...
  /api/notifications/me:
    get:
      summary: Consultar notificaciones
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Notificaciones y contador no leído }